test(translationMacro): enable, modernize, TDD auto-lookup
Tests need a `-test` suffix. This test was never working before.
